Expose Top level API through zombienet-sdk crate (#126)
Following up on take 1
https://github.com/paritytech/zombienet-sdk/pull/124

Apart from better logs, I realize that one of the only thing missing now
for us to start consuming this project is a top level package available
(with it's dependencies) on crates.io. This PR serves mainly as a
discussion point to get there.
It would be great if we could publish an early version (0.0.1?) on
crates.io so that we can start using this asap.