feat: enhance public api (#209)
Firstly, thank you for an amazing library! The original zombienet is a
Polkadot staple and now being able to use it completely from Rust is a
game changer for us: as you may be aware, we have integrated
zombienet-sdk with
[pop-cli](https://github.com/paritytech/zombienet-sdk).

In doing so, we needed to add a few getters to the public api to be able
to present certain information to the user, as well we expose a little
more of the inner errors. This PR simply upstreams these changes.

It also includes two additional fixes, which may be better implemented
based on your guidance:
- collator args werent working for us when specified in a config file,
which seemed to be due to missing `--` in the args list provided by
calling functions.
- artifact path wasnt set when trying to build a chain spec. This may no
longer be required based on the changes in
https://github.com/paritytech/zombienet-sdk/commit/c7ce50c0e54c7efd505d18c0d77fe169f7df0465

I am happy to add tests if you are open to these changes. 🙂